Refactor run_flo into modules
